chore : 악기 좌표 Integer로 변경

+++ backend/letsnote/src/main/java/com/geeks/letsnote/domain/studio/workSpace/dto/RequestNotes.java
@@ -7,7 +7,7 @@ public record RequestNotes() {
     @Builder
     public record NoteDto(
             Instrument instrument,
-            Long noteX,
-            Long noteY
+            Integer noteX,
+            Integer noteY
     ){}
 }
